Don't try to skip CI when force-closing the PR
There is just no nice way to do this,
https://github.com/orgs/community/discussions/102876
would be the best approach, but it is still not supported.

`gh workflow run` doesn't seem to really support
running a workflow on the PR itself.
Note that the `validate-<...>` job runs by `pull_request_target`,
and that runs not on the pr branch...